+++ src/main/java/com/parcel/parcelcosting/service/ParcelService.java
@@ -2,8 +2,6 @@ package com.parcel.parcelcosting.service;
 
 import com.parcel.parcelcosting.entity.Parcel;
 
-import java.math.BigDecimal;
-
 public interface ParcelService {
-    public BigDecimal getCost(Parcel parcel);
+    public Double getCost(Parcel parcel);
 }
