+++ src/main/java/com/parcel/parcelcosting/service/ParcelServiceImpl.java
@@ -6,14 +6,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 
 @Service
 public class ParcelServiceImpl implements ParcelService {
     Logger logger = LoggerFactory.getLogger(ParcelService.class);
     @Override
-    public BigDecimal getCost(Parcel parcel){
+    public Double getCost(Parcel parcel){
         return ParcelFactory.getDeliveryCost(parcel);
 
     }
